revert to original uint8 and uint32 gadgets

+++ dpc/src/dpc/base_dpc/program_circuit.rs
@@ -9,7 +9,7 @@ use snarkos_models::{
     gadgets::{
         algorithms::{CRHGadget, CommitmentGadget},
         r1cs::{ConstraintSynthesizer, ConstraintSystem},
-        utilities::{alloc::AllocGadget, uint::UInt8},
+        utilities::{alloc::AllocGadget, uint8::UInt8},
     },
 };
