+++ gadgets/src/curves/templates/bls12/affine.rs
@@ -13,7 +13,7 @@ use snarkos_models::{
             boolean::Boolean,
             eq::{ConditionalEqGadget, EqGadget, NEqGadget},
             select::CondSelectGadget,
-            uint::UInt8,
+            uint8::UInt8,
             ToBitsGadget,
             ToBytesGadget,
         },
